feat: add activity overlay, google auth
The activity overlay is for when a user doesn't have the app active. It's a
preliminary component for helping us determine when to fetch google data later.
The login/auth required a lot of stuff in the Google Cloud Console, but now
that it's all set up, it does kinda seem to work like magic...

[daily-write]